backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.44.0
scipy>=1.12.0
groq>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
